created sound mixer and re-organized files
