+++ Assets/Prefabs/Sounds/SoundEffectTemplate.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 2f7a7edc6502fd241b43262a6a996867
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
