+++ Assets/Scripts/Environment/Sound.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 63ef4eb961808664988b516f2de02348
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
