+++ Assets/Scripts/Environment/Sound/SoundEffectLibrary.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 91893622292ef7c4c82e8bbb390a67b0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
