+++ Assets/Scripts/Environment/Sound/SoundEffectManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f68b942f80e18244ea72896cf03c34aa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
