+++ Assets/Scripts/Environment/Sound/SoundEffectOnHit.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c08297e4b3940974bbd0e0aa886efc82
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
