+++ Assets/Sound/Settings.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c4c49b139f2828d40a1f67e21b9c84d2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
