+++ Assets/Sound/Settings/AudioMixer.mixer.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 447bcbade0387f642bb6a6663dd979ce
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 24100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
